💄 style(api/infra): rename crypto
