+++ apps/api/src/infrastructure/crypto/index.ts
