+++ apps/api/src/modules/auth-guard/tests/mobile.test.ts
@@ -1,7 +1,7 @@
 import { env } from "cloudflare:test";
 import { beforeEach, describe, expect, test } from "vitest";
 import { CLIENT_TYPE_HEADER_NAME } from "../../../common/constants";
-import { SessionSecretHasher } from "../../../infrastructure/crypt";
+import { SessionSecretHasher } from "../../../infrastructure/crypto";
 import { createSessionFixture, createUserFixture } from "../../../tests/fixtures";
 import { SessionTableHelper, UserTableHelper } from "../../../tests/helpers";
 import { ElysiaWithEnv } from "../../elysia-with-env";
