+++ apps/api/src/routes/auth/account-association/account-association-confirm.ts
@@ -3,7 +3,7 @@ import { AccountAssociationConfirmUseCase } from "../../../application/use-cases
 import { ValidateAccountAssociationSessionUseCase } from "../../../application/use-cases/account-association";
 import { ACCOUNT_ASSOCIATION_SESSION_COOKIE_NAME, SESSION_COOKIE_NAME } from "../../../common/constants";
 import { newAccountAssociationSessionToken } from "../../../domain/value-objects";
-import { SessionSecretHasher } from "../../../infrastructure/crypt";
+import { SessionSecretHasher } from "../../../infrastructure/crypto";
 import { DrizzleService } from "../../../infrastructure/drizzle";
 import { AccountAssociationSessionRepository } from "../../../interface-adapter/repositories/account-association-session";
 import { ExternalIdentityRepository } from "../../../interface-adapter/repositories/external-identity";
