+++ apps/api/src/routes/auth/login.ts
@@ -1,7 +1,7 @@
 import { t } from "elysia";
 import { LoginUseCase } from "../../application/use-cases/auth";
 import { SESSION_COOKIE_NAME } from "../../common/constants";
-import { PasswordHasher, SessionSecretHasher } from "../../infrastructure/crypt";
+import { PasswordHasher, SessionSecretHasher } from "../../infrastructure/crypto";
 import { DrizzleService } from "../../infrastructure/drizzle";
 import { SessionRepository } from "../../interface-adapter/repositories/session";
 import { UserRepository } from "../../interface-adapter/repositories/user";
