+++ apps/api/src/routes/me/update-email.ts
@@ -5,7 +5,7 @@ import {
 } from "../../application/use-cases/email-verification";
 import { EMAIL_VERIFICATION_SESSION_COOKIE_NAME, SESSION_COOKIE_NAME } from "../../common/constants";
 import { newEmailVerificationSessionToken } from "../../domain/value-objects";
-import { SessionSecretHasher } from "../../infrastructure/crypt";
+import { SessionSecretHasher } from "../../infrastructure/crypto";
 import { DrizzleService } from "../../infrastructure/drizzle";
 import { EmailVerificationSessionRepository } from "../../interface-adapter/repositories/email-verification-session";
 import { SessionRepository } from "../../interface-adapter/repositories/session";
